+++ packages/cubejs-jdbc-driver/driver/JDBCDriver.js
@@ -56,6 +56,34 @@ const DbTypes = {
       PWD: process.env.CUBEJS_AWS_SECRET,
       S3OutputLocation: process.env.CUBEJS_AWS_S3_OUTPUT_LOCATION
     }
+  },
+  sparksql: {
+    driverClass: "org.apache.hive.jdbc.HiveDriver",
+    prepareConnectionQueries: [],
+    mavenDependency: {
+      "groupId": "org.apache.hive",
+      "artifactId": "hive-jdbc",
+      "version": "2.3.5"
+    },
+    jdbcUrl: () => `jdbc:hive2://${process.env.CUBEJS_DB_HOST}:${process.env.CUBEJS_DB_PORT || "10000"}/${process.env.CUBEJS_DB_NAME}`,
+    properties: {
+      user: process.env.CUBEJS_DB_USER,
+      password: process.env.CUBEJS_DB_PASS,
+    }
+  },
+  hive: {
+    driverClass: "org.apache.hive.jdbc.HiveDriver",
+    prepareConnectionQueries: [],
+    mavenDependency: {
+      "groupId": "org.apache.hive",
+      "artifactId": "hive-jdbc",
+      "version": "2.3.5"
+    },
+    jdbcUrl: () => `jdbc:hive2://${process.env.CUBEJS_DB_HOST}:${process.env.CUBEJS_DB_PORT || "10000"}/${process.env.CUBEJS_DB_NAME}`,
+    properties: {
+      user: process.env.CUBEJS_DB_USER,
+      password: process.env.CUBEJS_DB_PASS,
+    }
   }
 };